added overflowing content detection on triggered in-view
This is used to indicate if the element that is brought into view has
overflowing content in it, e.g., elements that uses **overflow:
auto.**,  used especially for text contents.

Sort of having an in-view within an in-view, albeit the inner view is
detected in advanced if some of its content is partially
cropped/hidden, so the app can visually indicate (with ellipsis for
example) to users that the content they are seeing has information that
is cropped/hidden by the overflow: auto.

This pure CSS text-ellipsis works on one line only:
http://stackoverflow.com/questions/14183887/how-to-make-text-overflow-el
lipsis-work-on-floating-divs

Following pure CSS works on multi-line. Perfect on Safari, works on
both overflown width and height. On Chrome, text ellipsis works on
overflown height only. Does not work on other browsers:

http://codepen.io/martinwolf/pen/qlFdp